backend/post_service/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
pydantic==1.10.7
motor==3.1.2
pymongo==4.3.3
httpx==0.24.0
python-multipart==0.0.6
pyjwt==2.6.0
python-dotenv==1.0.0
email-validator==2.0.0
